+++ ORKA/UserDefinedApis/CellLookUpTable/cellLookupTable.h
@@ -12,6 +12,6 @@
 #include <stdint.h>
 
 CellTable_ts AE_tableFindByCapacity(uint32_t capacity, uint8_t MOLICEL_TABLE_);
-CellTable_ts AE_tableFindByVoltage(float voltage, uint8_t MOLICEL_TABLE_);
+CellTable_ts AE_tableFindByVoltage(double voltage, uint8_t MOLICEL_TABLE_);
 
 #endif /* ORKA_USERDEFINEDAPIS_CELLLOOKUPTABLE_CELLLOOKUPTABLE_H_ */
